sdk/pyproject.toml
[tool.poetry]
name = "chatfaq_sdk"
version = "0.1.0"
description = ""
authors = ["Hector Soria <attosoria@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
asyncio = "^3.4.3"
websockets = "^10.4"
coloredlogs = "^15.0.1"
sphinx = "<6"
myst-parser = "^0.18.1"
isort = "^5.11.4"
black = "^22.12.0"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
